pylint; fixes according to pylint
