+++ implementation/merge.py
@@ -1,13 +1,13 @@
-from .logger import printline, printlog
-from .dimension import Dimension
-from .files import is_image_extension, is_pdf_extension, is_document_extension
-from pathlib import Path
-from typing import Sequence
-from .configuration import Configuration
-import pymupdf
 import subprocess
 import tempfile
 import os
+from pathlib import Path
+from typing import Sequence
+import pymupdf
+from .configuration import Configuration
+from .logger import printline, printlog
+from .dimension import Dimension
+from .files import is_image_extension, is_pdf_extension, is_document_extension
 
 PathLike = str | Path
 
@@ -19,7 +19,9 @@ def libre_to_pdf(document_path: Path, config: Configuration, output_file: pymupd
         return
     tempdir = Path(tempfile.gettempdir()).joinpath("Zszywacz")
     os.makedirs(tempdir, exist_ok=True)
-    subprocess.run([config.libreoffice_path, "--convert-to", "pdf", str(document_path), "--outdir", tempdir])
+    subprocess.run(
+        [config.libreoffice_path, "--convert-to", "pdf", str(document_path), "--outdir", tempdir], check=True
+    )
     output_file.insert_file(
         tempdir.joinpath(document_path.with_suffix(".pdf").name)
     )  # insert_file can handle pathlib.Path
